libpod: move linux-specific code to _linux.go
This fixes a few "unused" linter warnings on freebsd.
