+++ libpod/container_inspect.go
@@ -11,7 +11,6 @@ import (
 	"github.com/containers/podman/v5/libpod/driver"
 	"github.com/containers/podman/v5/pkg/signal"
 	"github.com/containers/podman/v5/pkg/util"
-	"github.com/containers/storage/types"
 	"github.com/docker/go-units"
 	spec "github.com/opencontainers/runtime-spec/specs-go"
 	"github.com/sirupsen/logrus"
@@ -491,17 +490,6 @@ func (c *Container) generateInspectContainerConfig(spec *spec.Spec) *define.Insp
 	return ctrConfig
 }
 
-func generateIDMappings(idMappings types.IDMappingOptions) *define.InspectIDMappings {
-	var inspectMappings define.InspectIDMappings
-	for _, uid := range idMappings.UIDMap {
-		inspectMappings.UIDMap = append(inspectMappings.UIDMap, fmt.Sprintf("%d:%d:%d", uid.ContainerID, uid.HostID, uid.Size))
-	}
-	for _, gid := range idMappings.GIDMap {
-		inspectMappings.GIDMap = append(inspectMappings.GIDMap, fmt.Sprintf("%d:%d:%d", gid.ContainerID, gid.HostID, gid.Size))
-	}
-	return &inspectMappings
-}
-
 // Generate the InspectContainerHostConfig struct for the HostConfig field of
 // Inspect.
 func (c *Container) generateInspectContainerHostConfig(ctrSpec *spec.Spec, namedVolumes []*ContainerNamedVolume, mounts []spec.Mount) (*define.InspectContainerHostConfig, error) {
@@ -657,29 +645,6 @@ func (c *Container) generateInspectContainerHostConfig(ctrSpec *spec.Spec, named
 	hostConfig.ConsoleSize = []uint{0, 0}
 
 	return hostConfig, nil
-}
-
-// Return true if the container is running in the host's PID NS.
-func (c *Container) inHostPidNS() (bool, error) {
-	if c.config.PIDNsCtr != "" {
-		return false, nil
-	}
-	ctrSpec, err := c.specFromState()
-	if err != nil {
-		return false, err
-	}
-	if ctrSpec.Linux != nil {
-		// Locate the spec's PID namespace.
-		// If there is none, it's pid=host.
-		// If there is one and it has a path, it's "ns:".
-		// If there is no path, it's default - the empty string.
-		for _, ns := range ctrSpec.Linux.Namespaces {
-			if ns.Type == spec.PIDNamespace {
-				return false, nil
-			}
-		}
-	}
-	return true, nil
 }
 
 func (c *Container) GetDevices(priv bool, ctrSpec spec.Spec, deviceNodes map[string]string) ([]define.InspectDevice, error) {
